refactor: vo 클래스 패키지 이동
- 불필요한 import문 삭제
