+++ src/test/java/woowacourse/auth/domain/customer/vo/NicknameTest.java
@@ -1,8 +1,7 @@
-package woowacourse.auth.domain;
+package woowacourse.auth.domain.customer.vo;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
-import static org.junit.jupiter.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
