+++ src/test/java/woowacourse/auth/domain/customer/vo/PhoneNumberTest.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.domain;
+package woowacourse.auth.domain.customer.vo;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
